styling fixes on homepage and removed unnecessary http request code from login form and register form
