+++ client/components/IntroText/introText.module.scss
@@ -5,6 +5,5 @@
     flex-direction: column;
     align-items: center;
     color: v.$neutral-500;
-    font-size: v.$font-size-lg;
     gap: .5rem;
 }
