+++ client/components/LoginAndRegisterHandler/LoginAndRegisterHandler.tsx
@@ -15,8 +15,8 @@ function LoginAndRegisterHandler() {
             priority
             src="/images/right-arrow.svg"
             alt="arrow"
-            width={70}
-            height={70}
+            width={50}
+            height={50}
           />
         </div>
       </Link>
