+++ client/styles/App.module.scss
@@ -4,6 +4,6 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    gap: 5rem;
+    gap: 3rem;
     margin-top: 3rem;
 }
