+++ client/styles/variables.module.scss
@@ -12,7 +12,7 @@ $neutral-700: #4A524B;
 $neutral-900: #141916;
 
 // font-sizes
-$base-font-size: 1rem;
+$base-font-size: 13px;
 $font-size-sm: $base-font-size * 0.75;
 $font-size-lg: $base-font-size * 1.5;
 $font-size-xl: $base-font-size * 2;
